refactor(fs): consolidate the filesystem subsystem under fl/fs/
Filesystem code was spread across four unrelated parts of the tree with
no directory naming the subsystem: the API in fl/system/ alongside delay,
pin, heap and serial; `filebuf` — the base class every backend subclasses
— in fl/stl/detail/ because it is spelled like std::filebuf; the SD
backend in fl/system/sd/; and platform fragments in five places.

Tracing how a backend worked meant leaving the filesystem code to find
its own base class. Adding a second backend meant deciding where each
file went, repeatedly, with no precedent to follow.

    src/fl/fs/
        fs.h            FileSystem + medium factories
        fs.cpp.hpp
        backend.h       FsImpl — the platform extension point
        read.h          format helpers as free functions
        file_handle.h   filebuf      (from fl/stl/detail/)
        fstream.h       ifstream     (from fl/stl/)
        sd/             SD backend

`FsImpl` moving to its own header is the point of the split: sketch
authors and platform implementers previously read one file for unrelated
reasons.

`read.h` adds free functions over `FileSystem&` — readJpeg, readMp3,
readVideo, readJson, readScreenMap and the rest — forwarding to the
existing members, which stay for source compatibility. Adding a format
then costs one function there rather than widening FileSystem, which
today carries a member per format (#4003). Inverting the delegation and
dropping the members is a follow-up, since that is the breaking half.

`getSdFs(cs_pin)` joins `make_sdcard_filesystem` under a name describing
the medium, matching getEmbeddedFs() for on-chip flash so no driver name
reaches the public API (#4007).

fl/system/file_system.h becomes a forwarding header, so all 27 existing
includers — and any sketch in the wild — keep compiling unchanged.

Per-backend translation units are preserved: fl.fs.sd+.cpp still carries
the SD chain on its own so the linker drops it when beginSd() is never
called (#2773 item 1.2).

The noexcept ratchet baseline is path-rewritten rather than regenerated.
Regenerating locally produced 6344 entries against 7093, because this
toolchain parses fewer translation units than whoever generated it —
committing that would have silently weakened the ratchet for everyone.

bash lint clean, bash test --cpp passes, and FxSdCard still builds for
WASM.

Closes #4008
Refs #4003, #4007, #3986
