+++ ci/lint_cpp_rs/src/checkers/unity_build.rs
@@ -43,7 +43,8 @@ const UNITY_EXPECTED_BUILD_FILES: &[&str] = &[
     "fl/build/fl.sensors+.cpp",
     "fl/build/fl.stl+.cpp",
     "fl/build/fl.system+.cpp",
-    "fl/build/fl.system.sd+.cpp",
+    "fl/build/fl.fs+.cpp",
+    "fl/build/fl.fs.sd+.cpp",
     "fl/build/fl.task+.cpp",
     "fl/build/fl.test+.cpp",
     "fl/build/fl.ui+.cpp",
