+++ src/fl/fs/sd/_build.cpp.hpp
@@ -1,4 +1,4 @@
-/// @file fl/system/sd/_build.cpp.hpp
+/// @file fl/fs/sd/_build.cpp.hpp
 /// @brief Unity-build aggregate for SD-card support. Pulled in only by
 /// `src/fl/build/fl.system.sd+.cpp` — never by the parent
 /// `src/fl/system/_build.cpp.hpp`. That separation is what lets the
@@ -6,8 +6,8 @@
 /// VFSImpl, ~16 KB on ESP32-S3) when the user never calls
 /// `FileSystem::beginSd()`. See FastLED #2773 item 1.2.
 
-#include "fl/system/sd/file_system_sd.cpp.hpp"
-#include "fl/system/sd/fs_sdcard_arduino.cpp.hpp"
+#include "fl/fs/sd/file_system_sd.cpp.hpp"
+#include "fl/fs/sd/fs_sdcard_arduino.cpp.hpp"
 
 #if defined(FL_IS_TEENSY)
 #define FASTLED_TEENSY_SDFAT_IMPL "platforms/arm/teensy/sdfat/common/_build.cpp.hpp"
