+++ src/fl/stl/detail/_build.cpp.hpp
@@ -2,5 +2,4 @@
 /// @brief Unity build header for fl/stl/detail/ directory
 /// Includes all implementation files in alphabetical order
 
-#include "fl/stl/detail/file_handle.cpp.hpp"
 #include "fl/stl/detail/string_holder.cpp.hpp"
