+++ src/fl/system/_build.cpp.hpp
@@ -4,7 +4,6 @@
 #include "fl/system/delay.cpp.hpp"
 #include "fl/system/engine_events.cpp.hpp"
 #include "fl/system/fastled_internal.cpp.hpp"
-#include "fl/system/file_system.cpp.hpp"
 #include "fl/system/heap.cpp.hpp"
 #include "fl/system/pin.cpp.hpp"
 #include "fl/system/pins.cpp.hpp"
