+++ src/fl/video/pixel_stream.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "fl/system/file_system.h"
+#include "fl/fs/fs.h"
 #include "fl/stl/shared_ptr.h"         // For FASTLED_SHARED_PTR macros
 #include "fl/stl/int.h"
 #include "fl/stl/noexcept.h"
