+++ tests/fl/audio/detector/vocal_real_audio.hpp
@@ -7,7 +7,7 @@
 #include "fl/audio/audio_context.h"
 #include "fl/audio/detector/vocal.h"
 #include "fl/codec/vorbis.h"
-#include "fl/stl/detail/file_handle.h"
+#include "fl/fs/file_handle.h"
 #include "fl/stl/detail/file_io.h"
 #include "fl/math/math.h"
 #include "fl/stl/scope_exit.h"
