+++ tests/fl/codec/gif.hpp
@@ -1,5 +1,5 @@
 #include "test.h"
-#include "fl/system/file_system.h"
+#include "fl/fs/fs.h"
 #include "fl/codec/gif.h"
 #include "fl/fx/frame.h"
 #include "fl/stl/detail/memory_file_handle.h"
