+++ tests/fl/codec/jpeg.hpp
@@ -1,5 +1,5 @@
 #include "test.h"
-#include "fl/system/file_system.h"
+#include "fl/fs/fs.h"
 #include "fl/codec/jpeg.h"
 #include "fl/stl/detail/memory_file_handle.h"
 #include "fl/fx/frame.h"
