+++ tests/fl/codec/mp4_parser.hpp
@@ -1,6 +1,6 @@
 #include "test.h"
 #include "fl/codec/mp4_parser.h"
-#include "fl/system/file_system.h"
+#include "fl/fs/fs.h"
 #include "platforms/stub/fs_stub.hpp" // ok platform headers
 
 static fl::FileSystem setupCodecFilesystem_mp4() {
