+++ tests/fl/codec/mpeg1.hpp
@@ -1,5 +1,5 @@
 #include "test.h"
-#include "fl/system/file_system.h"
+#include "fl/fs/fs.h"
 #include "fl/codec/mpeg1.h"
 #include "fl/fx/frame.h"
 #include "fl/stl/detail/memory_file_handle.h"
