+++ tests/fl/fs/file_handle.cpp
@@ -1,5 +1,5 @@
 #include "test.h"
-#include "fl/stl/detail/file_handle.h"
+#include "fl/fs/file_handle.h"
 #include "fl/stl/cerrno.h"
 #include "platforms/stub/fs_stub.hpp"
