+++ tests/fl/fs/fs.cpp
@@ -19,7 +19,7 @@
 
 #include "test.h"
 
-#include "fl/system/file_system.h"
+#include "fl/fs/fs.h"
 #ifdef FASTLED_TESTING
 #include "platforms/stub/fs_stub.hpp"
 #include "fl/stl/stdint.h"
