+++ tests/fl/fx/video.cpp
@@ -12,7 +12,7 @@
 #include "fl/stl/new.h"
 #include "fl/stl/vector.h"
 #include "test.h"
-#include "fl/system/file_system.h"
+#include "fl/fs/fs.h"
 #include "fl/gfx/crgb.h"
 #include "fl/fx/fx.h"
 #include "fl/fx/fx2d.h"
