+++ tests/profile/json_performance.cpp
@@ -8,7 +8,7 @@
 //   ./json_performance large        # Run large JSON benchmark (requires benchmark_1mb.json)
 //   ./json_performance all          # Run all benchmarks
 
-#include "fl/system/file_system.h"
+#include "fl/fs/fs.h"
 #include "fl/stl/int.h"
 #include "fl/stl/json.h"
 #include "fl/stl/chrono.h"
